+++ pulsar/apps/wsgi/wrappers.py
@@ -507,7 +507,7 @@ class WsgiRequest(EnvironMixin):
         :class:`~.MultiValueDict` containing data from
         the request body, and data from uploaded files.
 
-        If the body data is not ready, return a :class:`.Future`
+        If the body data is not ready, return a :class:`~asyncio.Future`
         which results in the tuple.
 
         The result is cached.