+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Ver. 0.8.4 - Development
 * Several bug fixes in wsgi :class:`.Router`
 * Added :attr:`.Route.name` attribute
 * :class:`.WsgiResponse` does not send cookies back to the client by default
+* Critical bug fix for multiprocessing sockets when running on python 3.4 and
+  windows
 
 Ver. 0.8.3 - 2014-Jun-23
 ===========================