+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-Ver. 0.8.4 - Development
+Ver. 0.8.4 - 2014-Jul-07
 ===========================
 * Several bug fixes in wsgi :class:`.Router`
 * Added :attr:`.Route.name` attribute